Change search param from route param to query param
