+++ app/search/SearchResultGroup.tsx
